17 MVD/entrebagles/RIPE Atlas/requirements.txt
flask
flask-jsonpify
flask-sqlalchemy
flask-restful
dnspython==1.15.0
